+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/Options.java
@@ -3,8 +3,7 @@ package com.pichillilorenzo.flutter_inappwebview;
 import java.util.Map;
 
 public interface Options<T> {
-  static String LOG_TAG = "Options";
   public Options parse(Map<String, Object> options);
   public Map<String, Object> toMap();
-  public Map<String, Object> getRealOptions(T webView);
+  public Map<String, Object> getRealOptions(T obj);
 }