+++ example/lib/in_app_webiew_example.screen.dart
@@ -23,6 +23,16 @@ class _InAppWebViewExampleScreenState extends State<InAppWebViewExampleScreen> {
     super.initState();
 
     contextMenu = ContextMenu(
+      menuItems: [
+        ContextMenuItem(androidId: 1, iosId: "1", title: "Special", action: () async {
+          print("Menu item Special clicked!");
+          print(await webView.getSelectedText());
+          await webView.clearFocus();
+        })
+      ],
+      options: ContextMenuOptions(
+        hideDefaultSystemContextMenuItems: true
+      ),
       onCreateContextMenu: (hitTestResult) async {
         print("onCreateContextMenu");
         print(hitTestResult.extra);
@@ -31,16 +41,11 @@ class _InAppWebViewExampleScreenState extends State<InAppWebViewExampleScreen> {
       onHideContextMenu: () {
         print("onHideContextMenu");
       },
-      onContextMenuActionItemClicked: (contextMenuItemClicked) {
+      onContextMenuActionItemClicked: (contextMenuItemClicked) async {
         var id = (Platform.isAndroid) ? contextMenuItemClicked.androidId : contextMenuItemClicked.iosId;
         print("onContextMenuActionItemClicked: " + id.toString() + " " + contextMenuItemClicked.title);
       }
     );
-    contextMenu.menuItems = [
-      ContextMenuItem(androidId: 1, iosId: "1", title: "Special", action: () async {
-        print("Menu item Special clicked!");
-      })
-    ];
   }
 
   @override
@@ -74,17 +79,14 @@ class _InAppWebViewExampleScreenState extends State<InAppWebViewExampleScreen> {
                   BoxDecoration(border: Border.all(color: Colors.blueAccent)),
                   child: InAppWebView(
                     contextMenu: contextMenu,
-                    // initialUrl: "https://github.com/flutter",
-                    initialFile: "assets/index.html",
+                    initialUrl: "https://github.com/flutter",
+                    // initialFile: "assets/index.html",
                     initialHeaders: {},
                     initialOptions: InAppWebViewGroupOptions(
                       crossPlatform: InAppWebViewOptions(
                         debuggingEnabled: true,
                         useShouldOverrideUrlLoading: true
                       ),
-                      android: AndroidInAppWebViewOptions(
-                        supportMultipleWindows: true
-                      )
                     ),
                     onWebViewCreated: (InAppWebViewController controller) {
                       webView = controller;