+++ ios/Classes/ContextMenuOptions.swift
@@ -1,0 +1,17 @@
+//
+//  ContextMenuOptions.swift
+//  flutter_inappwebview
+//
+//  Created by Lorenzo Pichilli on 30/05/2020.
+//
+
+import Foundation
+
+class ContextMenuOptions: Options<NSObject> {
+    
+    var hideDefaultSystemContextMenuItems = false;
+
+    override init(){
+        super.init()
+    }
+}